Split monte_carlo into operations and utils
Refactor monte_carlo by extracting operation implementations and helpers from config.py into new modules. Added operations.py (simulate_dgp, raw_data_datagen, run_* tests/regression/filter) and operation_utils.py (helper routines like _clone_or_pass_shocks, _resolve_context_array, _select_raw_rep_array). Moved MCReferenceConstruct into reference_constructs.py. Updated __init__.py and config.py to import and re-export the relocated symbols (using internal _-prefixed imports where appropriate) and adjusted MCStep factory functions to reference the new callables. Aims to improve modularity and separate concerns; no intended behavioral changes.
